# docs/shop-api-openapi.yaml
...
openapi: 3.0.0
info:
  title: shop-api
  description: The final project for "Web Development with Java and Spring" course
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /users/login:
    put:
      tags:
        - User
      summary: log in and assign cart
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "200":
          description: Ok. User logged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionInfo'
        "403":
          description: Forbidden. Wrong credentials.
  /users/register:
    post:
      tags:
        - User
      summary: register new user
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "201":
          description: Created. New user created.
        "400":
          description: Bad Request. Wrong input.
        "409":
          description: Conflict. User with this email already exists.
  /users/reset:
    patch:
      tags:
        - User
      summary: reset password 1 step
      operationId: requestPasswordReset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "200":
          description: Ok. Reset token sent.
        "404":
          description: Not Found. User not found.
    put:
      tags:
        - User
      summary: reset password 2 step
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "200":
          description: Ok. New password saved.
        "400":
          description: Bad Request. Wrong input.
        "404":
          description: Not Found. User not found.
components:
  schemas:
...
    SessionInfo:
      type: object
      readOnly: true
      properties:
        sessionId:
          type: string
          format: token
          readOnly: true
          example: "b58b270b-e7a4-441a-a912-a982885b3ee6"
        userId:
          type: integer
          format: int64
          readOnly: true
          example: 4534
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 4533
        email:
          type: string
          format: email
          example: "my@email.com"
        password:
          type: string
          format: password
          writeOnly: true
          pattern: "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[!@#$%^&*()])[A-Za-z\\d!@#$%^&*()]{5,}$"
          description: "Password must be at least 5 characters long, with at least one letter, one digit, and one special character"
          example: "234dfs&"
...